requirements.txt
pipecat-ai[daily,cartesia,openai,silero,deepgram]
fastapi==0.115.6
uvicorn
python-dotenv
python-multipart
aiohttp